+++ packages/terra-slide-panel/src/terra-dev-site/doc/example/DefaultSlidePanel.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
+/* eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 import SlidePanel from 'terra-slide-panel/lib/SlidePanel';
 import classNames from 'classnames/bind';
 import styles from './DefaultSlidePanel.module.scss';