+++ packages/terra-time-input/src/terra-dev-site/test/time-input/twelve-hour/Second.test.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import TimeInput from '../../../../TimeInput';
 import TimeUtil from '../../../../TimeUtil';
+import styles from '../time-input/time-input.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 class TimeInputDefault extends React.Component {
   constructor(props) {
@@ -22,7 +26,7 @@ Time Input:
             {this.state.input}
           </h3>
         </div>
-        <div style={{ caretColor: 'transparent' }}>
+        <div className={cx(['content-wrapper'])}>
           <TimeInput
             id="timeInput"
             name="time-input"